requirements.txt
asgiref==3.6.0
Django==4.2
django-phonenumber-field==7.0.2
phonenumbers==8.13.10
sqlparse==0.4.4
